[pre-push] Fix bug when pushing empty stack
Previously, an empty stack would result in a PR's URL being derived from
a call to `gh` which would produce no output, and thus would crash
gherrit. We move that check logic later in the function to a point where
`gh` is guaranteed to produce output since gherrit will have returned by
that point for an empty stack.

gherrit-pr-id: G8e4c6bccdddd28364fb6f9e237a2c0454b6a2325