Replace branch "privacy" with pushRemote config
Instead of tracking whether a branch is "private" (as a GHerrit
setting), set `pushRemote = "."` when running `gherrit manage` (and
unset it when running `gherrit unmanage`). In the `pre-push` hook,
unconditionally exit 0. `pushRemote = "."` will have the effect of `git
push` pushing to "this" local repo, which is an effect-free no-op. This
allows us to remove the warning text about intentionally making `git
push` fail.

Closes #93

gherrit-pr-id: G582bbdae9bb1e48729f68075315d8d21f7200a48