[pre-push] Batch ls-remote
Previously, in order to avoid exceeding the maximum number of
command-line arguments to `git ls-remote`, we would fall back to
querying for *all* remote GHerrit refs when the number of command-line
args would otherwise exceed a certain threshold. In this commit, we
change this behavior to simply batch and make repeated calls to `git
ls-remote` if necessary. This should reduce the memory and bandwidth
required to download potentially large numbers of refs.

gherrit-pr-id: G164545d27727b8fc2d6266d00fa1b8b0d5a6dced