[pre-push] Don't prune merged commits
Previously, the `pre-push` hook did a `git fetch` operation on
`origin/main` in order to determine which commits had already been
merged upstream, and pruned these from the list of pushed commits. This
turned out to be fundamentally broken, resulting in in PRs with multiple
commits and with merge conflicts relative to `main`.

We could *theoretically* perform a more intelligent pruning operation
that *also* auto-rebases commits on the fly, but even if we managed to
do this without corrupting local state (since the naive way of doing
rebasing changes the local work tree, `HEAD` pointer, etc), the
resulting state would be opaque and confusing for the user.

Instead, we opt for the simpler approach of just letting the user see
that the push has failed and suggesting that they rebase on the latest
upstream changes – this is much more reliable, and it's a flow that
developers are already familiar with.

gherrit-pr-id: Gafe3a719b0d63cde7a135bcf2836afa4c865edd3