[pre-push] Prevent clobbering others' changes
When pushing the version tag to GitHub, use `--force-with-lease=<ref>:`
to instruct GitHub to reject version tags which already exist. This has
the effect of causing the push to fail if another developer or machine
has already pushed this version. More precisely, it ensures that pushes
only succeed when the local machine is completely up-to-date with
GitHub.

Closes #139

gherrit-pr-id: G54cf1d0f51eca8d64e6b9f22d572ddf289a66000