[pre-push] Replace `gh` CLI with `octocrab` crate
Use `octocrab`'s ability to issue GraphQL requests to batch PR creation
and update operations, allowing us to issue a single API request for
each in the common case (currently, if there are fewer than 50 PRs in a
stack). Note that this does *not* guarantee atomicity – batched
operations are still executed sequentially and independently on the
server.

In order to support testing this change, we need to support overriding
the default GitHub API URL. This is dangerous to allow in production, as
an attacker who can inject environment variables could cause a victim's
`gherrit` to connect to an attacker-controlled URL and provide the
victim's real GitHub API token. In order to mitigate this risk, this
commit also adds support for the `gherrit` binary to determine whether
it was built in a testing environment (specifically, if
`GHERRIT_TEST_BUILD` is set). It only allows overriding the GitHub API
URL in testing, but not in production.

gherrit-pr-id: G53274dc7071f4fc3b6494b801d49833927faccf0