[pre-push] Replace `gh` CLI with `octocrab` crate
Use `octocrab`'s ability to issue GraphQL requests to batch PR creation
and update operations, allowing us to issue a single API request for
each in the common case (currently, if there are fewer than 50 PRs in a
stack). Note that this does *not* guarantee atomicity – batched
operations are still executed sequentially and independently on the
server.

gherrit-pr-id: G53274dc7071f4fc3b6494b801d49833927faccf0