[pre-push] Shrink size of patch comparison table
Use a variety of techniques to shrink the size (in bytes) of the patch
comparison table in the generated Markdown of the PR body:
- If the total body size would exceed half of GitHub's limit, only
  render comparisons:
  - With the most recent version
  - With any version and:
    - The base branch
    - That version's previous version
- Remove non-rendered spaces from the Markdown table
- Use relative links, stripping the `https://github.com` prefix

This makes progress on #244, but does not fully solve the problem – it's
still possible for a PR with a very large number of versions to exceed
GitHub's limit and thus be un-updatable. A rough back-of-the-envelope
calculation suggests that somewhere between 500 and 2,000 patches would
be enough to make a PR un-updatable.

gherrit-pr-id: G2cc1c781a5c1407d19e702b4f11774bc0c83226e